requirement.txt
google-generativeai 
streamlit 
beautifulsoup4 
python-dotenv